backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.40
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.2
pymongo==4.10.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Machine Learning & AI
tensorflow==2.18.0
scikit-learn==1.6.1
opencv-python==4.11.0.86
pillow==11.1.0
numpy==2.2.2
pandas==2.2.3

# Weather & Geospatial
requests==2.32.3
geopy==2.4.1
geoalchemy2==0.16.1

# Communication
twilio==9.4.0
sendgrid==6.11.0

# File Storage & Processing
boto3==1.36.6
python-magic==0.4.27

# Background Tasks
celery==5.4.0
flower==2.0.1

# Monitoring & Logging
structlog==24.4.0
sentry-sdk[fastapi]==2.22.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
ruff==0.9.2

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.51

# Environment
python-dotenv==1.0.1